move walls_to_graph from utils.py to team.py
